requirements.txt
# Core Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# HTTP Client for LLM API calls
httpx>=0.26.0

tenacity>=8.2.0

# Environment variable management
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

black>=23.0.0
isort>=5.13.0
mypy>=1.8.0
